상품 조회 수 , 상품 조회 Redis 캐시 구현

+++ src/main/java/com/soloproject/shoppingmall/product/dto/ProductPatchDto.java
@@ -6,10 +6,9 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ProductPatchDto {
-
     private long productId;
     private String name;
     private String description;
-    private long price;
-    private long stock;
+    private int price;
+    private int stock;
 }
